backend/requirements.txt
google-cloud-aiplatform
functions-framework==3.*
Flask
Flask-Cors
numpy
gunicorn